removed get_data for BigNumber
