+++ superset/assets/src/visualizations/BigNumber/index.js
@@ -1,0 +1,5 @@
+import adapter from './adaptor';
+import BigNumber from './BigNumber';
+
+export { BigNumber };
+export default adapter;
